fix/moved features to own files
